requirements.txt
pycryptodomex
pyyaml==5.1
psycopg2-binary
boto3